+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Add liveness and readiness checks to Mirror Maker
 * Allow configuring loadBalancerIP for LoadBalancer type services
 * Allow setting labels and annotations for Persistent Volume Claims
+* Add support for Jaeger tracing in Kafka Mirror Maker and Kafka Connect
 
 ## 0.13.0